Error handling and extract notification client
