+++ packages/liveblocks-core/src/client.ts
@@ -2,22 +2,18 @@ import type { AuthValue } from "./auth-manager";
 import { createAuthManager } from "./auth-manager";
 import { isIdle } from "./connection";
 import { DEFAULT_BASE_URL } from "./constants";
-import {
-  convertToInboxNotificationData,
-  convertToThreadData,
-} from "./convert-plain-data";
 import type { LsonObject } from "./crdts/Lson";
 import { linkDevTools, setupDevTools, unlinkDevTools } from "./devtools";
 import { kInternal } from "./internal";
 import type { BatchStore } from "./lib/batch";
-import { Batch, createBatchStore } from "./lib/batch";
+import { createBatchStore } from "./lib/batch";
 import { createStore, Store } from "./lib/create-store";
 import { deprecateIf } from "./lib/deprecation";
 import * as console from "./lib/fancy-console";
 import type { Json, JsonObject } from "./lib/Json";
 import type { Resolve } from "./lib/Resolve";
+import { createInboxNotificationsApi } from "./notifications";
 import type { CustomAuthenticationResult } from "./protocol/Authentication";
-import { TokenKind } from "./protocol/AuthToken";
 import type { BaseUserMeta } from "./protocol/BaseUserMeta";
 import type { Polyfills, Room, RoomDelegates, RoomInitializers } from "./room";
 import {
@@ -28,12 +24,9 @@ import {
 import type { CacheStore } from "./store";
 import { createClientStore } from "./store";
 import type { BaseMetadata } from "./types/BaseMetadata";
-import type {
-  InboxNotificationData,
-  InboxNotificationDataPlain,
-} from "./types/InboxNotificationData";
+import type { InboxNotificationData } from "./types/InboxNotificationData";
 import type { OptionalPromise } from "./types/OptionalPromise";
-import type { ThreadData, ThreadDataPlain } from "./types/ThreadData";
+import type { ThreadData } from "./types/ThreadData";
 
 const MIN_THROTTLE = 16;
 const MAX_THROTTLE = 1_000;
@@ -46,7 +39,6 @@ const MAX_LOST_CONNECTION_TIMEOUT = 30_000;
 const DEFAULT_LOST_CONNECTION_TIMEOUT = 5_000;
 
 const RESOLVE_USERS_BATCH_DELAY = 50;
-const MARK_INBOX_NOTIFICATIONS_AS_READ_BATCH_DELAY = 50;
 
 export type ResolveMentionSuggestionsArgs = {
   /**
@@ -100,7 +92,9 @@ type PrivateClientApi<TUserMeta extends BaseUserMeta> = {
   usersStore: BatchStore<TUserMeta["info"] | undefined, [string]>;
 };
 
-type InboxNotificationsApi<TThreadMetadata extends BaseMetadata = never> = {
+export type InboxNotificationsApi<
+  TThreadMetadata extends BaseMetadata = never,
+> = {
   /**
    * @private
    */
@@ -283,10 +277,6 @@ export type ClientOptions<TUserMeta extends BaseUserMeta = BaseUserMeta> = {
 //     | ((room: string) => Promise<{ token: string }>);
 //
 
-export type GetInboxNotificationsOptions = {
-  limit?: number;
-};
-
 function getBaseUrlFromClientOptions(clientOptions: ClientOptions) {
   if ("liveblocksServer" in clientOptions) {
     throw new Error("Client option no longer supported");
@@ -357,8 +347,6 @@ export function createClient<TUserMeta extends BaseUserMeta = BaseUserMeta>(
   };
 
   const roomsById = new Map<string, RoomInfo>();
-
-  const notificationsApi = createInboxNotificationsApi(fetchClientApi);
 
   function teardownRoom(room: OpaqueRoom) {
     unlinkDevTools(room.id);
@@ -529,33 +517,17 @@ export function createClient<TUserMeta extends BaseUserMeta = BaseUserMeta>(
 
   const currentUserIdStore = createStore<string | null>(null);
 
-  async function fetchClientApi(
-    endpoint: string,
-    options?: RequestInit
-  ): Promise<Response> {
-    const authValue = await authManager.getAuthValue();
-
-    if (
-      authValue.type !== "secret" ||
-      authValue.token.parsed.k === TokenKind.SECRET_LEGACY
-    ) {
-      throw new Error("TODO");
-    }
-
-    const userId = authValue.token.parsed.uid;
-    currentUserIdStore.set(() => userId);
-
-    const url = new URL(`/v2/c${endpoint}`, baseUrl);
-    const fetcher =
-      clientOptions.polyfills?.fetch || /* istanbul ignore next */ fetch;
-    return await fetcher(url.toString(), {
-      ...options,
-      headers: {
-        ...options?.headers,
-        Authorization: `Bearer ${getAuthBearerHeaderFromAuthValue(authValue)}`,
-      },
-    });
-  }
+  const {
+    getInboxNotifications,
+    getUnreadInboxNotificationsCount,
+    markAllInboxNotificationsAsRead,
+    markInboxNotificationAsRead,
+  } = createInboxNotificationsApi({
+    baseUrl,
+    fetcher: clientOptions.polyfills?.fetch || /* istanbul ignore next */ fetch,
+    authManager,
+    currentUserIdStore,
+  });
 
   const cacheStore = createClientStore();
 
@@ -597,7 +569,10 @@ export function createClient<TUserMeta extends BaseUserMeta = BaseUserMeta>(
       enterRoom,
 
       // Notifications API
-      ...notificationsApi,
+      getInboxNotifications,
+      getUnreadInboxNotificationsCount,
+      markAllInboxNotificationsAsRead,
+      markInboxNotificationAsRead,
 
       // Internal
       [kInternal]: {
@@ -622,116 +597,6 @@ export class NotificationsApiError extends Error {
   ) {
     super(message);
   }
-}
-
-function createInboxNotificationsApi(
-  fetchClientApi: (endpoint: string, options?: RequestInit) => Promise<Response>
-): InboxNotificationsApi {
-  async function fetchJson<T>(
-    endpoint: string,
-    options?: RequestInit
-  ): Promise<T> {
-    const response = await fetchClientApi(endpoint, options);
-
-    if (!response.ok) {
-      if (response.status >= 400 && response.status < 600) {
-        let error: NotificationsApiError;
-
-        try {
-          const errorBody = (await response.json()) as { message: string };
-
-          error = new NotificationsApiError(
-            errorBody.message,
-            response.status,
-            errorBody
-          );
-        } catch {
-          error = new NotificationsApiError(
-            response.statusText,
-            response.status
-          );
-        }
-
-        throw error;
-      }
-    }
-
-    let body;
-
-    try {
-      body = (await response.json()) as T;
-    } catch {
-      body = {} as T;
-    }
-
-    return body;
-  }
-
-  async function getInboxNotifications(options?: GetInboxNotificationsOptions) {
-    const queryParams = toURLSearchParams({ limit: options?.limit });
-    const json = await fetchJson<{
-      // [comments-unread] TODO: How do we type ThreadMetadata?
-      threads: ThreadDataPlain[];
-      inboxNotifications: InboxNotificationDataPlain[];
-    }>(`/inbox-notifications?${queryParams.toString()}`);
-
-    return {
-      threads: json.threads.map((thread) => convertToThreadData(thread)),
-      inboxNotifications: json.inboxNotifications.map((notification) =>
-        convertToInboxNotificationData(notification)
-      ),
-    };
-  }
-
-  async function getUnreadInboxNotificationsCount() {
-    const { count } = await fetchJson<{
-      count: number;
-    }>("/inbox-notifications/count");
-
-    return count;
-  }
-
-  async function markAllInboxNotificationsAsRead() {
-    await fetchJson("/inbox-notifications/read", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ inboxNotificationIds: "all" }),
-    });
-  }
-
-  async function markInboxNotificationsAsRead(inboxNotificationIds: string[]) {
-    await fetchJson("/inbox-notifications/read", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ inboxNotificationIds }),
-    });
-  }
-
-  const batchedMarkInboxNotificationsAsRead = new Batch(
-    async (batchedInboxNotificationIds: [string][]) => {
-      const inboxNotificationIds = batchedInboxNotificationIds.flat();
-
-      await markInboxNotificationsAsRead(inboxNotificationIds);
-
-      return inboxNotificationIds;
-    },
-    { delay: MARK_INBOX_NOTIFICATIONS_AS_READ_BATCH_DELAY }
-  );
-
-  async function markInboxNotificationAsRead(inboxNotificationId: string) {
-    await batchedMarkInboxNotificationsAsRead.get(inboxNotificationId);
-  }
-
-  return {
-    getInboxNotifications,
-    getUnreadInboxNotificationsCount,
-    markAllInboxNotificationsAsRead,
-    markInboxNotificationAsRead,
-  };
 }
 
 function checkBounds(
@@ -778,35 +643,4 @@ function getLostConnectionTimeout(value: number): number {
     MAX_LOST_CONNECTION_TIMEOUT,
     RECOMMENDED_MIN_LOST_CONNECTION_TIMEOUT
   );
-}
-
-/**
- * Safely but conveniently build a URLSearchParams instance from a given
- * dictionary of values. For example:
- *
- *   {
- *     "foo": "bar+qux/baz",
- *     "empty": "",
- *     "n": 42,
- *     "nope": undefined,
- *     "alsonope": null,
- *   }
- *
- * Will produce a value that will get serialized as
- * `foo=bar%2Bqux%2Fbaz&empty=&n=42`.
- *
- * Notice how the number is converted to its string representation
- * automatically and the `null`/`undefined` values simply don't end up in the
- * URL.
- */
-function toURLSearchParams(
-  params: Record<string, string | number | null | undefined>
-): URLSearchParams {
-  const result = new URLSearchParams();
-  for (const [key, value] of Object.entries(params)) {
-    if (value !== undefined && value !== null) {
-      result.set(key, value.toString());
-    }
-  }
-  return result;
 }
